backend/requirements.txt
# API
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9

# NLP/ML
pandas==2.2.3
numpy==2.1.2
joblib==1.4.2
scikit-learn==1.5.2

# Transformers stack
transformers==4.44.2
sentence-transformers==3.0.1
torch>=2.2.0

# Resume parsing
python-docx==1.1.2
pdfplumber==0.11.4

# Excel (if you keep the services sheet as .xlsx)
openpyxl==3.1.5
